# openapi.yaml
openapi: 3.0.3
info:
  title: Zorvyn Finance API
  description: |
    Enterprise-grade Finance Data Processing and Access Control Backend.
    
    ## Authentication
    All protected endpoints require a Bearer token in the Authorization header.
    Obtain tokens via `POST /api/v1/auth/login`.
    
    ## Industry Features
    - **Correlation IDs**: Every request gets a unique `X-Correlation-ID` header
    - **Idempotency**: POST endpoints accept `X-Idempotency-Key` header to prevent duplicates
    - **Rate Limiting**: All endpoints are rate-limited (see `X-RateLimit-*` headers)
    - **Row-Level Scoping**: Non-admin users can only see data from their department
    - **Audit Trail**: Every mutation is logged with before/after snapshots
    - **Precision Money**: All amounts are NUMERIC(19,4) — passed as strings, never floats
  version: 1.0.0
  contact:
    name: Zorvyn Engineering
    email: engineering@zorvyn.io
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/users:
    get:
      tags: [Users]
      summary: List all users (paginated)
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: role
          in: query
          schema:
            type: string
            enum: [viewer, analyst, admin]
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive, suspended]
        - name: department
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of users
        '403':
          description: Admin only
...
  /api/v1/records:
    post:
      tags: [Records]
      summary: Create a financial record
      description: |
        Creates a new financial record. Supports idempotency via `X-Idempotency-Key` header.
        Amounts must be passed as strings (e.g., "15000.0000"). Admin only.
      security:
        - bearerAuth: []
      parameters:
        - name: X-Idempotency-Key
          in: header
          description: Unique key to prevent duplicate creation
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRecord'
      responses:
        '201':
          description: Record created
        '409':
          description: Duplicate request (idempotency key already used)
    get:
      tags: [Records]
      summary: List financial records (filtered, paginated, scoped)
      description: |
        Returns records scoped to the user's department (admins see all).
        Supports filtering by type, category, date range, and pagination.
      security:
        - bearerAuth: []
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [income, expense, transfer]
        - name: category
          in: query
          schema:
            type: string
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          schema:
            type: string
            format: date
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: sort
          in: query
          schema:
            type: string
            enum: [date, amount, created_at, category]
            default: date
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: Paginated list of records
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Obtain a token via POST /api/v1/auth/login.
        Use the accessToken in the Authorization header.
  schemas:
...
    CreateRecord:
      type: object
      required: [amount, type, category, date]
      properties:
        amount:
          type: string
          description: Decimal string — never a float (e.g., "15000.0000")
          example: "15000.0000"
        type:
          type: string
          enum: [income, expense, transfer]
          example: income
        category:
          type: string
          example: revenue
        description:
          type: string
          example: Monthly SaaS revenue
        date:
          type: string
          format: date
          example: "2026-03-15"
        department:
          type: string
          example: headquarters
...